+++ src/content/site-adapters/huggingchat-adapter.ts
@@ -1,0 +1,5 @@
+import { GenericAdapter } from './generic-adapter';
+
+export class HuggingChatAdapter extends GenericAdapter {
+  readonly name = 'HuggingChat';
+}