+++ src/content/site-adapters/perplexity-adapter.ts
@@ -1,0 +1,5 @@
+import { GenericAdapter } from './generic-adapter';
+
+export class PerplexityAdapter extends GenericAdapter {
+  readonly name = 'Perplexity';
+}