+++ src/content/site-adapters/poe-adapter.ts
@@ -1,0 +1,5 @@
+import { GenericAdapter } from './generic-adapter';
+
+export class PoeAdapter extends GenericAdapter {
+  readonly name = 'Poe';
+}